load statistics: Fix a bunch of integer overflow bugs
Most of the summed up load statistics were skewed by a factor of 10, due
to a series of integer multiplications which was overflowing the 32-bit
result holder.

load, load1m, load10m before this patch (16 processes, 3 fully loaded):
12 99 9
12 99 9
12 101 10
13 103 10
13 104 10
12 106 10

... and after:

12 9 0
11 10 1
12 10 1
13 10 1

Fixes #1650

+++ pt_load.c
@@ -298,7 +298,7 @@ unsigned int pt_get_1m_load(int _)
 			summed_procs++;
 		}
 
-	return (used*100/(LT_WINDOW_TIME*summed_procs*LT_1m_RATIO));
+	return (used*100/((long long)LT_WINDOW_TIME*summed_procs*LT_1m_RATIO));
 }
 
 
@@ -320,7 +320,7 @@ unsigned int pt_get_10m_load(int _)
 			summed_procs++;
 		}
 
-	return (used*100/(LT_WINDOW_TIME*summed_procs));
+	return (used*100/((long long)LT_WINDOW_TIME*summed_procs));
 }
 
 
@@ -341,7 +341,7 @@ unsigned int pt_get_rt_loadall(int _)
 			summed_procs++;
 		}
 
-	return (used*100/(ST_WINDOW_TIME*summed_procs));
+	return (used*100/((long long)ST_WINDOW_TIME*summed_procs));
 }
 
 
@@ -362,7 +362,7 @@ unsigned int pt_get_1m_loadall(int _)
 			summed_procs++;
 		}
 
-	return (used*100/(LT_WINDOW_TIME*summed_procs*LT_1m_RATIO));
+	return (used*100/((long long)LT_WINDOW_TIME*summed_procs*LT_1m_RATIO));
 }
 
 
@@ -383,7 +383,7 @@ unsigned int pt_get_10m_loadall(int _)
 			summed_procs++;
 		}
 
-	return (used*100/(LT_WINDOW_TIME*summed_procs));
+	return (used*100/((long long)LT_WINDOW_TIME*summed_procs));
 }
